Update JMeter test and API specification
Updated the JMeter test, SmokeTest.jmx, replacing ResponseAssertions with HTMLAssertions and increasing the UniformRandomTimer delay. In the API specification, the 'text' parameter is replaced with 'sentence' for more clarity, and new response body descriptions are added. Also, several new Rapidoc and Redoc HTML files have been added to the resources.