Remove httpYac workflow, update API tests and unit tests workflows
The httpYac API tests workflow was removed due to redundancy and maintenance efficiency. Meanwhile, adjustments were made to the API tests and unit tests workflow files to ensure that they are functioning properly. This includes modifying the publish test report sections and adding a step to publish unit test results.