Update endpoints in translation counter
The endpoint names in the translation counter have been updated from '/actuator/translation-count' to '/actuator/count'. This change was made to simplify the naming conventions in the code. No functionality has been affected and applications using these endpoints should update the URL they are using to engage with this feature.